added the optionals.

+++ app/Bootstrap.java
@@ -1,11 +1,15 @@
-import play.jobs.*;
+import play.jobs.Job;
+import play.jobs.OnApplicationStart;
 import play.test.Fixtures;
+import models.*;
 
 @OnApplicationStart
 public class Bootstrap extends Job {
 	public void doJob() {
+		if (Survey.count()==0) {
 		// Load default data if the database is empty
 		Fixtures.deleteDatabase();
 		Fixtures.loadModels("data.yml");
+		}
 	}
 }
